a52dec: add -fPIC to CFLAGS
Without -fPIC gst-plugins.ugly fails to compile. The problem
is that a52dec builds a static library. This library is then used
to create a gstreamer-plugin .so. The linking fails if a52dec is
compiled without -fPIC.

This problem manifests itself only when hardening is disabled.
As when hardening is enabled -fPIC is added automatically.